[WIFI-1545] Moving CustomerPortalDashboardPartialEvent to status-models
