+++ status-models/src/test/java/com/telecominfraproject/wlan/status/dashboard/models/CustomerPortalDashboardTests.java
@@ -8,8 +8,8 @@ import org.slf4j.LoggerFactory;
 
 import com.telecominfraproject.wlan.core.model.equipment.RadioType;
 import com.telecominfraproject.wlan.core.model.json.BaseJsonModel;
+import com.telecominfraproject.wlan.status.dashboard.models.events.CustomerPortalDashboardPartialEvent;
 import com.telecominfraproject.wlan.status.models.StatusCode;
-import com.telecominfraproject.wlan.systemevent.aggregation.models.CustomerPortalDashboardPartialEvent;
 
 public class CustomerPortalDashboardTests {
 
@@ -23,12 +23,12 @@ public class CustomerPortalDashboardTests {
 		evtPartial.incrementAssociatedClientsCountPerRadio(RadioType.is5GHzU, 3);
 		evtPartial.incrementClientCountPerOui("000000", 8);
 		evtPartial.incrementClientCountPerOui("0F0F0F", 22);
-		evtPartial.incrementAlarmsCountBySeverity(StatusCode.requiresAttention.name(), 10);
+		evtPartial.incrementAlarmsCountBySeverity(StatusCode.requiresAttention, 10);
 		
 		CustomerPortalDashboardStatus status = new CustomerPortalDashboardStatus();
 		status.applyPartialEvent(evtPartial);
 		status.incrementEquipmentCountPerOui("2F2F2F", 14);
-		status.incrementAlarmsCountBySeverity(StatusCode.requiresAttention.name(), 10);
+		status.incrementAlarmsCountBySeverity(StatusCode.requiresAttention, 10);
 		
 		String statusStr = status.toString();
 		
@@ -46,7 +46,7 @@ public class CustomerPortalDashboardTests {
 		evt.incrementAssociatedClientsCountPerRadio(RadioType.is5GHzU, 3);
 		evt.incrementClientCountPerOui("000000", 8);
 		evt.incrementClientCountPerOui("0F0F0F", 22);
-		evt.incrementAlarmsCountBySeverity(StatusCode.requiresAttention.name(), 10);
+		evt.incrementAlarmsCountBySeverity(StatusCode.requiresAttention, 10);
 		
 		String evtStr = evt.toString();
